Add libspdlog v1.5.0 compatibility
The API should compile against the library installed in the system
and not the one coming from Hunter dependency system.
The reason is that ROS2 also uses libspdlog but older version.
Both libraries should be aligned and in this case it is easier to
make DepthAI API to use older version of the library.